+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.2.0] - 2024-05-20
+## [1.2.1] - 2026-04-23
+
+### ⚠️ Breaking Changes
+- **onOverscrolled**: The `onOverscrolled` method now receives `finished: Boolean, direction: Direction` instead of `finished: Boolean`. 
+
+### 🛠 Bug Fixes
+- **Direction mix-up**: `Direction` enum is correctly calculated for the `OverscrolledEffectNode`
+
+## [1.2.0] - 2026-04-23
 
 ### ⚠️ Breaking Changes
 - **OverscrolledEffectNode**: The `node` method now receives `currentProgress: () -> OverscrolledProgress` instead of `currentOffset: () -> Offset`.
@@ -12,7 +20,7 @@ All notable changes to this project will be documented in this file.
 ### 🚀 Features
 - **onProgressChanged**: Added a new optional callback to `rememberHorizonalOverscrolledEffect` and `rememberVerticalOverscrolledEffect` to observe overscroll progress changes.
 
-## [1.1.0] - 2023-10-27
+## [1.1.0] - 2026-04-22
 
 ### 🚀 Features
 - **Asymmetrical Thresholds**: Added support for separate start and end overscroll thresholds in `rememberHorizonalOverscrolledEffect` and `rememberVerticalOverscrolledEffect`.